Add defer option
When you're animating things, you need to know the the position
at the beginning and the position at the end of the animation.
Easy right ?

The problem is that when you're animating stuff, it's highly likely
that your dom elements are not existing beforehand. Thus, when
getting the start position, you miss the dom element since it does
not exist yet.

The defer option added to the ReactFlipContainer allows you to defer
execution of the animation. Technically that means that you've got
a round of React update to mount your elements and put them in the
correct initial style. You can take this opportunity to measure how
big your dom element is. Or more basically set the opacity at 0 so that
the resulting animation would be a a fade in. Or both!

However there is still one problem left yet : you can't really combine
defer animations with standard ones. I'm working on it in order to fix
this.

Plus the element used in the toggle demo could be extracted into the core
library, making things easier to reuse. But first consolidate the inner
details before doing so.